pyproject.toml
[project]
name = "image-generation-bot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiogram>=3.23.0",
    "google-api-python-client>=2.150.0",
    "google-auth>=2.36.0",
    "google-auth-oauthlib>=1.2.0",
    "google-genai>=1.0.0",
    "gspread>=6.1.0",
    "httpx>=0.28.1",
    "loguru>=0.7.3",
    "pillow>=10.0.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.0.0",
    "python-slugify>=8.0.4",
    "sqlalchemy>=2.0.45",
]
